+++ mod-source-record-storage-server/src/main/java/org/folio/rest/impl/SourceStorageSourceRecordsImpl.java
@@ -7,18 +7,19 @@ import static org.folio.dao.util.RecordDaoUtil.filterRecordByLeaderRecordStatus;
 import static org.folio.dao.util.RecordDaoUtil.filterRecordByRecordId;
 import static org.folio.dao.util.RecordDaoUtil.filterRecordBySnapshotId;
 import static org.folio.dao.util.RecordDaoUtil.filterRecordBySuppressFromDiscovery;
-import static org.folio.dao.util.RecordDaoUtil.filterRecordByType;
 import static org.folio.dao.util.RecordDaoUtil.filterRecordByUpdatedDateRange;
 import static org.folio.dao.util.RecordDaoUtil.toRecordOrderFields;
+import static org.folio.rest.util.QueryParamUtil.toExternalIdType;
+import static org.folio.rest.util.QueryParamUtil.toRecordType;
 
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
+import java.util.UUID;
 
 import javax.ws.rs.NotFoundException;
 import javax.ws.rs.core.Response;
 
-import org.folio.dao.util.RecordType;
 import org.folio.dataimport.util.ExceptionHelper;
 import org.folio.rest.jaxrs.model.SourceRecord;
 import org.folio.rest.jaxrs.resource.SourceStorageSourceRecords;
@@ -58,21 +59,18 @@ public class SourceStorageSourceRecordsImpl implements SourceStorageSourceRecord
       Boolean suppressFromDiscovery, Boolean deleted, String leaderRecordStatus, Date updatedAfter, Date updatedBefore,
       List<String> orderBy, int offset, int limit, Map<String, String> okapiHeaders,
       Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
-    // NOTE: if and when a new record type is introduced and a parsed record table is added,
-    // will need to add a record type query parameter
     vertxContext.runOnContext(v -> {
       try {
         Condition condition = filterRecordByRecordId(recordId)
           .and(filterRecordBySnapshotId(snapshotId))
           .and(filterRecordByInstanceId(instanceId))
           .and(filterRecordByInstanceHrid(instanceHrid))
-          .and(filterRecordByType(recordType))
           .and(filterRecordBySuppressFromDiscovery(suppressFromDiscovery))
           .and(filterRecordByDeleted(deleted))
           .and(filterRecordByLeaderRecordStatus(leaderRecordStatus))
           .and(filterRecordByUpdatedDateRange(updatedAfter, updatedBefore));
           List<OrderField<?>> orderFields = toRecordOrderFields(orderBy, true);
-        recordService.getSourceRecords(condition, RecordType.valueOf(recordType), orderFields, offset, limit, tenantId)
+        recordService.getSourceRecords(condition, toRecordType(recordType), orderFields, offset, limit, tenantId)
           .map(GetSourceStorageSourceRecordsResponse::respond200WithApplicationJson)
           .map(Response.class::cast)
           .otherwise(ExceptionHelper::mapExceptionToResponse)
@@ -89,7 +87,7 @@ public class SourceStorageSourceRecordsImpl implements SourceStorageSourceRecord
       Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
     vertxContext.runOnContext(v -> {
       try {
-        recordService.getSourceRecords(entity, idType, RecordType.valueOf(recordType), deleted, tenantId)
+        recordService.getSourceRecords(entity, toExternalIdType(idType), toRecordType(recordType), deleted, tenantId)
           .map(GetSourceStorageSourceRecordsResponse::respond200WithApplicationJson)
           .map(Response.class::cast)
           .otherwise(ExceptionHelper::mapExceptionToResponse)
@@ -106,7 +104,7 @@ public class SourceStorageSourceRecordsImpl implements SourceStorageSourceRecord
       Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
     vertxContext.runOnContext(v -> {
       try {
-        recordService.getSourceRecordById(id, idType, tenantId)
+        recordService.getSourceRecordById(id, toExternalIdType(idType), tenantId)
           .map(optionalSourceRecord -> optionalSourceRecord.orElseThrow(() ->
             new NotFoundException(String.format(NOT_FOUND_MESSAGE, SourceRecord.class.getSimpleName(), id))))
           .map(GetSourceStorageSourceRecordsByIdResponse::respond200WithApplicationJson)