+++ mod-source-record-storage-server/src/main/java/org/folio/rest/util/QueryParamUtil.java
@@ -1,0 +1,47 @@
+package org.folio.rest.util;
+
+import javax.ws.rs.BadRequestException;
+
+import org.codehaus.plexus.util.StringUtils;
+import org.folio.dao.util.ExternalIdType;
+import org.folio.dao.util.RecordType;
+
+public class QueryParamUtil {
+
+  private QueryParamUtil() { }
+
+  /**
+   * Tries to convert string to {@link ExternalIdType}. Returns default RECORD if null or empty.
+   *
+   * @param externalIdType external id type as string
+   * @return external id type
+   */
+  public static ExternalIdType toExternalIdType(String externalIdType) {
+    if (StringUtils.isNotEmpty(externalIdType)) {
+      try {
+        return ExternalIdType.valueOf(externalIdType);
+      } catch (Exception e) {
+        throw new BadRequestException(String.format("Unknown id type %s", externalIdType));
+      }
+    }
+    return ExternalIdType.RECORD;
+  }
+
+  /**
+   * Tries to convert string to {@link RecordType}. Returns default MARC if null or empty.
+   *
+   * @param recordType record type as string
+   * @return record type
+   */
+  public static RecordType toRecordType(String recordType) {
+    if (StringUtils.isNotEmpty(recordType)) {
+      try {
+        return RecordType.valueOf(recordType);
+      } catch (Exception e) {
+        throw new BadRequestException(String.format("Unknown record type %s", recordType));
+      }
+    }
+    return RecordType.MARC;
+  }
+
+}