+++ mod-source-record-storage-server/src/main/java/org/folio/services/RecordService.java
@@ -4,6 +4,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
+import org.folio.dao.util.ExternalIdType;
 import org.folio.dao.util.RecordType;
 import org.folio.rest.jaxrs.model.ParsedRecordDto;
 import org.folio.rest.jaxrs.model.ParsedRecordsBatchResponse;
@@ -111,24 +112,24 @@ public interface RecordService {
   /**
    * Searches for {@link SourceRecord} where id in a list of ids defined by id type. i.e. INSTANCE or RECORD
    *
-   * @param ids        list of ids
-   * @param idType     id type
-   * @param recordType record type
-   * @param deleted    filter by state DELETED or leader record status d, s, or x
-   * @param tenantId   tenant id
+   * @param ids            list of ids
+   * @param externalIdType id type
+   * @param recordType     record type
+   * @param deleted        filter by state DELETED or leader record status d, s, or x
+   * @param tenantId       tenant id
    * @return future with {@link SourceRecordCollection}
    */
-  Future<SourceRecordCollection> getSourceRecords(List<String> ids, String idType, RecordType recordType, Boolean deleted, String tenantId);
+  Future<SourceRecordCollection> getSourceRecords(List<String> ids, ExternalIdType externalIdType, RecordType recordType, Boolean deleted, String tenantId);
 
   /**
    * Searches for source record by id via specific id type
    *
-   * @param id       for searching
-   * @param idType   search type
-   * @param tenantId tenant id
+   * @param id             for searching
+   * @param externalIdType search type
+   * @param tenantId       tenant id
    * @return future with optional source record
    */
-  Future<Optional<SourceRecord>> getSourceRecordById(String id, String idType, String tenantId);
+  Future<Optional<SourceRecord>> getSourceRecordById(String id, ExternalIdType externalIdType, String tenantId);
 
   /**
    * Update parsed records from collection of records and external relations ids in one transaction
@@ -142,23 +143,23 @@ public interface RecordService {
   /**
    * Searches for Record either by SRS id or external relation id
    *
-   * @param id       either SRS id or external relation id
-   * @param idType   specifies of external relation id type
-   * @param tenantId tenant id
+   * @param id             either SRS id or external relation id
+   * @param externalIdType specifies of external relation id type
+   * @param tenantId       tenant id
    * @return future with {@link Record}
    */
-  Future<Record> getFormattedRecord(String id, String idType, String tenantId);
+  Future<Record> getFormattedRecord(String id, ExternalIdType externalIdTypee, String tenantId);
 
   /**
    * Change suppress from discovery flag for record by external relation id
    *
-   * @param id       id
-   * @param idType   external id type
-   * @param suppress suppress from discovery
-   * @param tenantId tenant id
+   * @param id             id
+   * @param externalIdType external id type
+   * @param suppress       suppress from discovery
+   * @param tenantId       tenant id
    * @return future with true if succeeded
    */
-  Future<Boolean> updateSuppressFromDiscoveryForRecord(String id, String idType, Boolean suppress, String tenantId);
+  Future<Boolean> updateSuppressFromDiscoveryForRecord(String id, ExternalIdType externalIdType, Boolean suppress, String tenantId);
 
   /**
    * Deletes records associated with specified snapshot and snapshot itself