+++ mod-source-record-storage-server/src/test/java/org/folio/services/UpdatedRecordEventHandlingServiceTest.java
@@ -20,10 +20,16 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import org.folio.TestUtil;
 import org.folio.dao.RecordDao;
 import org.folio.dao.RecordDaoImpl;
+import org.folio.dao.util.ExternalIdType;
 import org.folio.dao.util.SnapshotDaoUtil;
 import org.folio.processing.events.utils.ZIPArchiver;
-import org.folio.rest.jaxrs.model.*;
+import org.folio.rest.jaxrs.model.ParsedRecord;
+import org.folio.rest.jaxrs.model.ParsedRecordDto;
+import org.folio.rest.jaxrs.model.RawRecord;
+import org.folio.rest.jaxrs.model.Record;
 import org.folio.rest.jaxrs.model.Record.State;
+import org.folio.rest.jaxrs.model.Snapshot;
+import org.folio.rest.jaxrs.model.SourceRecord;
 import org.folio.rest.jooq.Tables;
 import org.folio.rest.util.OkapiConnectionParams;
 import org.junit.After;
@@ -153,7 +159,7 @@ public class UpdatedRecordEventHandlingServiceTest extends AbstractLBServiceTest
       if (ar.failed()) {
         context.fail(ar.cause());
       }
-      recordService.getSourceRecordById(record.getMatchedId(), "RECORD", TENANT_ID).onComplete(getNew -> {
+      recordService.getSourceRecordById(record.getMatchedId(), ExternalIdType.RECORD, TENANT_ID).onComplete(getNew -> {
         if (getNew.failed()) {
           context.fail(getNew.cause());
         }